Various fixes to support uninstall
Cleaned up else branch for readability
Remove logic that checks if an application is synced before starting the
cascading deletion. The use case for removing a pattern, and
applications should support removing non healthy and non synced
applications as well.
Removed logic to update the app before deleting.
Removed the logic to check if ACM hub is present. We want to be able to
remove pattern and applications even if ACM hub was present before. ACM
hub removal preferably should happen on argocd and application level.